+++ packages/core/src/labels.ts
@@ -1,4 +1,4 @@
-import { Operation } from './operation';
+import { Operation, OperationObject } from './operation';
 import { isObjectOperation } from './predicates';
 import { Symbol } from './symbol';
 
@@ -44,4 +44,19 @@ export function setLabels<T>(operation: Operation<T>, labels: Labels): Operation
     operation.labels = labels;
   }
   return operation;
+}
+
+export function extractLabels<T>(operation: Operation<T>): Labels | undefined {
+  if (!isObjectOperation<T>(operation)) {
+    return operation?.labels;
+  }
+  let labels: Labels = {};
+  for (let key in operation) {
+    let value = operation[key];
+    if (typeof value == 'string'
+    || typeof value == 'number'
+    || typeof value == 'boolean')
+      labels[key] = value;
+  }
+  return labels;
 }
