+++ packages/process/test/exec.test.ts
@@ -72,8 +72,8 @@ describe('exec', () => {
           let error: unknown;
           let proc = yield exec("argle", { arguments: ['bargle'] });
           try {
-            yield proc.expect();
-          } catch (e) { error = e }
+            yield proc.expect()
+          } catch (e) { error = e; }
 
           expect(error).toBeDefined();
         });
@@ -125,7 +125,7 @@ describe('exec', () => {
       });
 
       describe('when it fails', () => {
-        let error: Error;
+        let error: Error
         beforeEach(function*() {
           yield fetch('http://localhost:29000', { method: "POST", body: "fail" });
         });
@@ -140,10 +140,10 @@ describe('exec', () => {
             try {
               yield proc.expect();
             } catch (e) {
-              error = e as Error;
+              error = e;
             }
           };
-          expect(error).toBeDefined();
+          expect(error).toBeDefined()
         });
 
         it('closes stdout and stderr', function*() {
@@ -166,7 +166,7 @@ describe('exec', () => {
         let { stdout }: ProcessResult = yield proc.expect();
 
         expect(stdout).toEqual("second\n");
-      });
+      })
     });
   }
 
@@ -178,7 +178,7 @@ describe('exec', () => {
       let { stdout }: ProcessResult = yield proc.expect();
 
       expect(stdout).toEqual("first | echo second\n");
-    });
+    })
   });
 
 });
